requirements.txt
cs50
Flask
Flask-Session
requests
google-api-python-client